Make KuduSessionOptions immutable, and adjust some defaults
