+++ src/Knet.Kudu.Client/KuduSessionOptions.cs
@@ -1,62 +1,50 @@
 using System;
-using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Knet.Kudu.Client
 {
-    // TODO: Use C# 9 record class for this
-    public class KuduSessionOptions
+    public record KuduSessionOptions
     {
         /// <summary>
         /// The maximum number of items to send in a single write request.
         /// Hitting this number of items will trigger a flush immediately,
         /// regardless of the value of <see cref="FlushInterval"/>.
+        /// Default: 2000.
         /// </summary>
-        public int BatchSize { get; set; } = 1000;
+        public int BatchSize { get; init; } = 2000;
 
         /// <summary>
-        /// The maximum number of items the session may buffer.
+        /// The maximum number of items the session may store before calls to
+        /// <see cref="IKuduSession.EnqueueAsync(KuduOperation, CancellationToken)"/>
+        /// will block until a batch completes. Default: 40000.
         /// </summary>
-        public int Capacity { get; set; } = 100000;
+        public int Capacity { get; init; } = 40000;
 
         /// <summary>
         /// True if writers to the session guarantee that there will only ever
         /// be at most one write operation at a time; false if no such constraint
-        /// is guaranteed.
+        /// is guaranteed. Default: false.
         /// </summary>
-        public bool SingleWriter { get; set; }
+        public bool SingleWriter { get; init; }
 
         /// <summary>
         /// The maximum duration of time to wait for new items before flushing.
         /// Hitting this duration will trigger a flush immediately, regardless
-        /// of the value of <see cref="BatchSize"/>.
+        /// of the value of <see cref="BatchSize"/>. Default: 1 second.
         /// </summary>
-        public TimeSpan FlushInterval { get; set; } = TimeSpan.FromSeconds(1);
+        public TimeSpan FlushInterval { get; init; } = TimeSpan.FromSeconds(1);
 
         /// <summary>
         /// The external consistency mode for this session.
+        /// Default: <see cref="ExternalConsistencyMode.ClientPropagated"/>.
         /// </summary>
-        public ExternalConsistencyMode ExternalConsistencyMode { get; set; } =
+        public ExternalConsistencyMode ExternalConsistencyMode { get; init; } =
             ExternalConsistencyMode.ClientPropagated;
 
         /// <summary>
         /// Optional callback to be notified of errors.
         /// </summary>
-        public Func<SessionExceptionContext, ValueTask> ExceptionHandler { get; set; }
-    }
-
-    public sealed class SessionExceptionContext
-    {
-        public Exception Exception { get; }
-
-        public IReadOnlyList<KuduOperation> Rows { get; }
-
-        public SessionExceptionContext(
-            Exception exception,
-            IReadOnlyList<KuduOperation> rows)
-        {
-            Exception = exception;
-            Rows = rows;
-        }
+        public Func<SessionExceptionContext, ValueTask> ExceptionHandler { get; init; }
     }
 }
